+++ src/ballistica/base/python/base_python.h
@@ -40,7 +40,7 @@ class BasePython {
     kTranslateCall,
     kLStrClass,
     kCallClass,
-    kGarbageCollectSessionEndCall,
+    kGarbageCollectCall,
     kConfig,
     kAppOnNativeBootstrappingCompleteCall,
     kResetToMainMenuCall,