+++ tools/efro/debug.py
@@ -91,9 +91,10 @@ def _getr(slist: list[Any], olist: list[Any], seen: set[int]) -> None:
 
 
 def _get_all_objects(expanded: bool) -> list[Any]:
-    """Return an expanded list of all objects.
+    """Return all objects visible to garbage collector.
 
-    See https://utcc.utoronto.ca/~cks/space/blog/python/GetAllObjects
+    For notes on the 'expanded' option, see:
+    https://utcc.utoronto.ca/~cks/space/blog/python/GetAllObjects
     """
     gcl = gc.get_objects()
     if not expanded:
@@ -110,16 +111,21 @@ def _get_all_objects(expanded: bool) -> list[Any]:
 
 
 def getobj(objid: int, expanded: bool = False) -> Any:
-    """Return a garbage-collected object by its id.
+    """Return a garbage-collectable object by its id.
 
     Remember that this is VERY inefficient and should only ever be used
     for debugging.
     """
+    # If they passed a string (hex, etc), convert to int.
+    # if isinstance(objid, str):
+    #     objid = int(objid, 0)  # Autodetect hex/etc.
+
     if not isinstance(objid, int):
         raise TypeError(f'Expected an int for objid; got a {type(objid)}.')
 
     # Don't wanna return stuff waiting to be garbage-collected.
-    gc.collect()
+    # UPDATE: Turning this off.
+    # gc.collect()
 
     allobjs = _get_all_objects(expanded=expanded)
     for obj in allobjs:
@@ -199,11 +205,21 @@ def printrefs(
         Can be a list of object ids; if that particular object is
         found, it will always be expanded even if max_level has been reached.
     """
+
+    # Let's always exclude the gc.garbage list. When we're debugging
+    # with gc.DEBUG_SAVEALL enabled this list will include everything,
+    # so this cuts out lots of noise.
+    if exclude_objs is None:
+        exclude_objs = []
+    else:
+        exclude_objs = list(exclude_objs)
+    exclude_objs.append(gc.garbage)
+
     _printrefs(
         obj,
         level=0,
         max_level=max_level,
-        exclude_objs=[] if exclude_objs is None else exclude_objs,
+        exclude_objs=exclude_objs,
         expand_ids=[] if expand_ids is None else expand_ids,
         file=sys.stderr if file is None else file,
     )
@@ -300,7 +316,7 @@ def _desc(obj: Any) -> str:
         # identify what.
         for ref in getrefs(obj):
             if hasattr(ref, '__dict__') and vars(ref) is obj:
-                extra = f' (vars for {_desctype(ref)} @ {id(ref)})'
+                extra = f' (vars for {_desctype(ref)} @ {hex(id(ref))})'
 
         # Generic dict: print length and the first few key:type pairs.
         if extra is None:
@@ -316,7 +332,7 @@ def _desc(obj: Any) -> str:
             extra = f' (len {len(obj)}{pairss})'
     if extra is None:
         extra = ''
-    return f'{_desctype(obj)} @ {id(obj)}{extra}'
+    return f'{_desctype(obj)} @ {hex(id(obj))}{extra}'
 
 
 def _printrefs(