+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-### 1.7.44 (build 22426, api 9, 2025-06-17)
+### 1.7.44 (build 22427, api 9, 2025-06-17)
 - Added a `-B` / `--dont-write-bytecode` flag to disable writing .pyc files, and
   an associated `dont_write_bytecode` value for the server config file. In most
   cases writing .pyc files is useful as it can speed up relaunches and keep
@@ -35,6 +35,15 @@
   them (makes it more consistent with hardware cursors).
 - Software cursor in sdl builds now disappears when the cursor leaves the window
   instead of getting stuck at the edge.
+- Replaced all uses of Python's built in `urllib.request` with our bundled
+  `urllib3`. This should perform better and hopefully won't get stuck at
+  shutdown like old urllib was prone to do. Please holler if you run into any
+  sort of connectivity issues that weren't there before.
+- Greatly cleaned up and improved `babase.garbage_collect()`. It will now warn
+  if too many objects are resorting to cyclic garbage collection due to
+  reference loops, and it offers some tips and functionality to help track down
+  and eliminate said loops. Flip the `ba.garbagecollection` log to `Debug` to
+  learn more.
 
 ### 1.7.43 (build 22406, api 9, 2025-06-09)
 - Fixes an issue with tournament scores not submitting properly in 1.7.42.